+++ src/org/uecide/gui/swing/SketchLibraryNode.java
@@ -25,13 +25,30 @@ import java.awt.Font;
 import java.awt.Component;
 import java.awt.BorderLayout;
 
-public class SketchLibraryNode extends SketchTreeNodeBase implements ContextEventListener {
+public class SketchLibraryNode extends SketchTreeNodeBase implements ContextEventListener, AnimationListener {
     Library library;
+
+    CleverIcon goodIcon = null;
+    CleverIcon badIcon = null;
+    CleverIcon emptyIcon = null;
+    CleverIcon compilingIcon = null;
 
     public SketchLibraryNode(Context c, SketchTreeModel m, Library l) {
         super(c, m, l.getName());
         library = l;
         updateChildren();
+        ctx.listenForEvent("libraryCompileStarted", this);
+        ctx.listenForEvent("libraryCompileFinished", this);
+        ctx.listenForEvent("libraryCompileFailed", this);
+
+        try {
+            goodIcon = IconManager.getIcon(16, "tree.lib-good");
+            badIcon = IconManager.getIcon(16, "tree.lib-bad");
+            emptyIcon = IconManager.getIcon(16, "tree.lib-bad");
+            compilingIcon = IconManager.getIcon(16, "main.spin");
+            compilingIcon.addAnimationListener(this);
+        } catch (Exception ex) {
+        }
     }
 
     public Library getLibrary() {
@@ -39,7 +56,19 @@ public class SketchLibraryNode extends SketchTreeNodeBase implements ContextEven
     }
 
     public ImageIcon getIcon(JTree tree) throws IOException {
-        return IconManager.getIcon(16, "library");
+        if (library.isCompiled()) {
+            return goodIcon;
+        }
+
+        if (library.isCompiling()) {
+            return compilingIcon;
+        }
+
+        if (library.compilingFailed()) {
+            return badIcon;
+        }
+
+        return emptyIcon;
     }
 
     public boolean updateChildren() {
@@ -55,9 +84,25 @@ public class SketchLibraryNode extends SketchTreeNodeBase implements ContextEven
     }
 
     public void contextEventTriggered(ContextEvent e) {
+        if (e.getObject() instanceof Library) {
+            Library lib = (Library)e.getObject();
+            if (lib == library) {
+                model.reload(this);
+            }
+        }
     }
 
     public Component getRenderComponent(JLabel original, JTree tree) {
+        try {
+            original.setIcon(getIcon(tree));
+        } catch (IOException ex) {
+        }
         return original;
+    }
+
+    public void animationUpdated(CleverIcon i) {
+        if (library.isCompiling()) {
+            model.reload(this);
+        }
     }
 }